+++ src/tests/repositories/activities/addActivityRepository.test.js
@@ -1,0 +1,12 @@
+
+const subjectUnderTest = require('@/repositories/activities/addActivityRepository').default;
+
+describe('add activity repository', () => {
+    beforeEach(() => {
+        jest.clearAllMocks();
+    });
+
+    test('empty activity item given', () => {
+        subjectUnderTest({});
+    });
+});